+++ website/application/models/dockservice.php
@@ -11,15 +11,21 @@ class DockService implements iService{
         }
     }
 
-    public function readHoldsBicycle($dock_id, $station_id){
-    	$stmt = $this->db->prepare("SELECT holds_bicycle FROM dock WHERE dock_id = ? AND station_id = ?");
-    	$stmt->bind_param("ii", $dock_id, $station_id);
-    	$stmt->execute();
-    	$stmt->bind_result($holds_bicycle);
-    	$stmt->fetch();
-    	$returnDock = new Dock($dock_id, $station_id, $holds_bicycle);
-    	$stmt->close();
-    	return $returnDock;
+    public function readHoldsBicycle($dock){
+        if(validate($dock))
+        {
+        	$stmt = $this->db->prepare("SELECT holds_bicycle FROM dock WHERE dock_id = ? AND station_id = ?");
+        	$stmt->bind_param("ii", $dock->dock_id, $dock->station_id);
+        	$stmt->execute();
+        	$stmt->bind_result($holds_bicycle);
+        	$stmt->fetch();
+        	$stmt->close();
+        	return new Dock($dock->dock_id, $dock->station_id, $holds_bicycle);
+        }
+        else
+        {
+            return null;
+        }
     }
 
     public function readAllDocksForStation($station_id){
@@ -62,31 +68,53 @@ class DockService implements iService{
     }
 
     public function create($dock){
-    	$stmt = $this->db->prepare("INSERT INTO dock(station_id) VALUES (?)");
-        $stmt->bind_param("i", $dock->station_id);
-        $stmt->execute();
-        $id = $this->db->insert_id;
-        $stmt->close();
+        if(validate($dock))
+        {
+        	$stmt = $this->db->prepare("INSERT INTO dock(station_id) VALUES (?)");
+            $stmt->bind_param("i", $dock->station_id);
+            $stmt->execute();
+            $id = $this->db->insert_id;
+            $stmt->close();
 
-        return new Dock($station_id, $id, null);
+            return new Dock($station_id, $id, null);
+        }
+        else
+        {
+            return null;
+        }
     }
 
     public function update($dock){
-    	$stmt = $this->db->prepare("UPDATE dock set station_id = ?, holds_bicycle = ? WHERE dock_id = ?");
-    	$stmt->bind_param("iii", $dock->station_id, $dock->holds_bicycle, $dock->dock_id);
-    	$stmt->execute();
-    	$stmt->close();
-    	return $dock;
+        if(validate($dock))
+        {
+        	$stmt = $this->db->prepare("UPDATE dock set station_id = ?, holds_bicycle = ? WHERE dock_id = ?");
+        	$stmt->bind_param("iii", $dock->station_id, $dock->holds_bicycle, $dock->dock_id);
+        	$stmt->execute();
+        	$stmt->close();
+        	return $dock;
+        }
+        else
+        {
+            return null;
+        }
     }
 
     public function delete($dock){
-    	$stmt = $this->db->prepare("DELETE FROM dock WHERE dock_id = ?, station_id = ?");
-    	$stmt->bind_param("ii", $dock->dock_id, $dock->station_id);
-    	$stmt->execute();
-    	$stmt->close();
+        if(validate($dock))
+        {
+        	$stmt = $this->db->prepare("DELETE FROM dock WHERE dock_id = ?, station_id = ?");
+        	$stmt->bind_param("ii", $dock->dock_id, $dock->station_id);
+        	$stmt->execute();
+        	$stmt->close();
+            return true;
+        }
+        else
+        {
+            return false;
+        }
     }
 
-    private function validate($dock)
+    public function validate($dock)
     {
         return true;
     }
